backend/requirements.txt
fastapi[all]
passlib[bcrypt]
SQLAlchemy
pyjwt
